+++ packages/tswebextension/src/cli/copyWar.ts
@@ -1,6 +1,6 @@
+/* eslint-disable no-console */
 import path from 'path';
 import { copy } from 'fs-extra';
-import { logger } from '../lib/common/utils/logger';
 
 const REDIRECTS_CONFIG_PATH = 'redirects.yml';
 const REDIRECTS_RESOURCES_SRC_PATH = 'redirect-files';
@@ -8,6 +8,7 @@ const REDIRECTS_RESOURCES_DEST_PATH = 'redirects';
 
 const src = path.resolve(require.resolve('@adguard/scriptlets'), '../..');
 
+// TODO: use logger from lib after import fix
 export const copyWar = async (dest: string): Promise<void> => {
     dest = path.resolve(process.cwd(), dest);
 
@@ -15,8 +16,8 @@ export const copyWar = async (dest: string): Promise<void> => {
         await copy(path.resolve(src, REDIRECTS_CONFIG_PATH), path.resolve(dest, REDIRECTS_CONFIG_PATH));
         await copy(path.resolve(src, REDIRECTS_RESOURCES_SRC_PATH), path.resolve(dest, REDIRECTS_RESOURCES_DEST_PATH));
 
-        logger.info(`Web accessible resources was copied to ${dest}`);
+        console.info(`Web accessible resources was copied to ${dest}`);
     } catch (e) {
-        logger.error((e as Error).message);
+        console.error((e as Error).message);
     }
 };